requirements.txt
altair==5.0.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
cryptography==41.0.5
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.1
django-cors-headers==4.3.0
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.4
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3

markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6

overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.10.24
pipreqs==0.4.13
platformdirs==3.11.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==1.7.1
Pympler==1.0.1
PyMySQL==1.1.0
pypandoc==1.11
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0

PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
Send2Trash==1.8.2
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2

tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.3.1
uri-template==1.2.0
urllib3==2.0.3
validators==0.20.0
virtualenv==20.24.5
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
widgetsnbextension==4.0.7
yarg==0.1.9
zipp==3.16.2
